Move registrate builders to `api.registry.registrate.builder`
