+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/builder/SingleblockMachineBuilder.java
@@ -1,10 +1,12 @@
-package com.gregtechceu.gtceu.api.registry.registrate;
+package com.gregtechceu.gtceu.api.registry.registrate.builder;
 
 import com.gregtechceu.gtceu.api.block.MetaMachineBlock;
 import com.gregtechceu.gtceu.api.item.MetaMachineItem;
 import com.gregtechceu.gtceu.api.machine.MachineDefinition;
 import com.gregtechceu.gtceu.api.machine.MachineInstanceFactory;
 import com.gregtechceu.gtceu.api.machine.MetaMachine;
+import com.gregtechceu.gtceu.api.registry.registrate.GTRegistrate;
+
 import com.tterrag.registrate.builders.BuilderCallback;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
