+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/builder/package-info.java
@@ -1,0 +1,4 @@
+@NotNullByDefault
+package com.gregtechceu.gtceu.api.registry.registrate.builder;
+
+import org.jetbrains.annotations.NotNullByDefault;
